added presistent ui of inventory bar and background music
