+++ Assets/pixel_assets/prefab/Inventory.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c210e783c5831254aa895803124ff8da
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
