+++ Assets/pixel_assets/prefab/LevelAwithPresist.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0db21bc75f0176042840d607182e8b29
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
