+++ Assets/pixel_assets/prefab/inventory slot.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8b229c53266361c448fc37ddf91766e6
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
