--- Assets/scripts/Player/player_controller.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4873f9edb3817b046a78acae2405aade
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
