+++ Assets/scripts/Sound.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ee20c3e95c18976449cbe1631450ea67
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
