+++ Assets/scripts/UI/ButtonLoad.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Persisting
+{
+    public class ButtonLoadLevel : MonoBehaviour
+    {
+        public void OnButtonLoadLevel(string levelName)
+        {
+            // Call the LoadLevel function through the static instance variable
+            LevelManagerPersist.instance.LoadLevel(levelName);
+        }
+    }
+}
