+++ Assets/scripts/UI/ButtonLoad.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ad0d6a4e1e3b9874b9334bf5dd70a9e4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
