+++ Assets/scripts/UI/UIManager.cs
@@ -1,0 +1,19 @@
+using UnityEngine;
+
+public class PersistentUI : MonoBehaviour
+{
+    private static PersistentUI instance;
+
+    void Awake()
+    {
+        if (instance == null)
+        {
+            instance = this;
+            DontDestroyOnLoad(gameObject); // Keeps this object between scenes
+        }
+        else
+        {
+            Destroy(gameObject); // Ensures only one instance exists
+        }
+    }
+}
